+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Fixed corner-case crash affecting the `pattern: $X` optimization ("empty And; no positive terms in And")
 - Added support for parsing labels and goto in PHP (#3592)
 - Added type inference for struct literals in Go (#3622)
+- Fix semgrep-core crash when a cache file exceeds the file size limit
 
 ## [0.64.0](https://github.com/returntocorp/semgrep/releases/tag/v0.64.0) - 09-01-2021